ID-486 delete old CreateUserSpec use new one
